LCD Display takes input, show h or l
LCD Display receives users' input, then sees whether or not it exceeds the range. If it does, it will indicate whether it is high or low in the final display. Program will not quit, user has unlimited chances of entering a new set of values. System goes by the assumption that both values can only be done at the same time.